# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Jikū API",
    "description" : "Event invitation, ticketing, RSVP and check-in API.",
    "version" : "v1"
  },
  "servers" : [
    {
      "url" : "http://localhost",
      "description" : "Generated server url"
    }
  ],
  "security" : [
    {
      "bearer-jwt" : [ ]
    }
  ],
  "paths" : {
...
    "/api/v1/prospects" : {
      "post" : {
        "tags" : [
          "prospect-lead-controller"
        ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProspectLeadRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProspectLeadAck"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{eventId}/payments/manual" : {
      "get" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "current",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "manual-payment-controller"
        ],
        "operationId" : "request",
        "parameters" : [
          {
            "name" : "eventId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ManualPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ManualPaymentInstructions"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/trials" : {
      "get" : {
        "tags" : [
          "admin-trial-controller"
        ],
        "operationId" : "list_3",
        "parameters" : [
          {
            "name" : "status",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "tenantId",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          {
            "name" : "page",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 0
            }
          },
          {
            "name" : "size",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 20
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AdminTrialView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "admin-trial-controller"
        ],
        "operationId" : "grant",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GrantTrialRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminTrialView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/trials/{id}/end" : {
      "post" : {
        "tags" : [
          "admin-trial-controller"
        ],
        "operationId" : "end",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EndTrialRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminTrialView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tenants/{id}/suspend" : {
      "post" : {
        "tags" : [
          "admin-tenant-controller"
        ],
        "operationId" : "suspend",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TenantStatusChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tenants/{id}/reactivate" : {
      "post" : {
        "tags" : [
          "admin-tenant-controller"
        ],
        "operationId" : "reactivate",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TenantStatusChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/prospects/{id}/contacted" : {
      "post" : {
        "tags" : [
          "admin-prospect-lead-controller"
        ],
        "operationId" : "markContacted",
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProspectLeadView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/tenants" : {
      "get" : {
        "tags" : [
          "admin-tenant-controller"
        ],
        "operationId" : "list_7",
        "parameters" : [
          {
            "name" : "query",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "page",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 0
            }
          },
          {
            "name" : "size",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 20
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantDirectoryPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/prospects" : {
      "get" : {
        "tags" : [
          "admin-prospect-lead-controller"
        ],
        "operationId" : "list_8",
        "parameters" : [
          {
            "name" : "status",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProspectLeadView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ProspectLeadRequest" : {
        "type" : "object",
        "properties" : {
          "businessName" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "contactName" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "phone" : {
            "type" : "string",
            "maxLength" : 32,
            "minLength" : 0
          },
          "sector" : {
            "type" : "string",
            "minLength" : 1
          },
          "email" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "email",
            "maxLength" : 255,
            "minLength" : 0
          },
          "city" : {
            "type" : [
              "string",
              "null"
            ],
            "maxLength" : 120,
            "minLength" : 0
          },
          "weeklyVolume" : {
            "type" : [
              "string",
              "null"
            ],
            "maxLength" : 20,
            "minLength" : 0
          },
          "note" : {
            "type" : [
              "string",
              "null"
            ],
            "maxLength" : 1000,
            "minLength" : 0
          },
          "source" : {
            "type" : [
              "string",
              "null"
            ],
            "maxLength" : 100,
            "minLength" : 0
          }
        },
        "required" : [
          "businessName",
          "contactName",
          "phone",
          "sector"
        ]
      },
      "ProspectLeadAck" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
      "ManualPaymentRequest" : {
        "type" : "object",
        "properties" : {
          "tier" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "tier"
        ]
      },
      "ManualPaymentInstructions" : {
        "type" : "object",
        "properties" : {
          "paymentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reference" : {
            "type" : "string"
          },
          "tier" : {
            "type" : "string"
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "payee" : {
            "$ref" : "#/components/schemas/PayeeDetails"
          }
        }
      },
      "PayeeDetails" : {
        "type" : "object",
        "properties" : {
          "payeeName" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyNumber" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "mobileMoneyOperator" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "bankDetails" : {
            "type" : [
              "string",
              "null"
            ]
          }
        }
      },
...
      "GrantTrialRequest" : {
        "type" : "object",
        "properties" : {
          "tenantId" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "uuid"
          },
          "eventId" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "uuid"
          },
          "tier" : {
            "type" : "string",
            "minLength" : 1
          },
          "expiresAt" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          }
        },
        "required" : [
          "eventId",
          "expiresAt",
          "tenantId",
          "tier"
        ]
      },
      "AdminTrialView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "tenantId" : {
            "type" : "string"
          },
          "eventId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "tier" : {
            "type" : "string"
          },
          "grantedAllowance" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "endedReason" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "EndTrialRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "reason"
        ]
      },
      "TenantStatusChangeRequest" : {
        "type" : "object",
        "properties" : {
          "note" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [
          "note"
        ]
      },
      "TenantInfo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "contactEmail" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "displayName" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "primaryColor" : {
            "type" : "string"
          },
          "legalIdentity" : {
            "oneOf" : [
              {
                "$ref" : "#/components/schemas/TenantLegalIdentityInfo"
              },
              {
                "type" : "null"
              }
            ]
          }
        }
      },
      "TenantLegalIdentityInfo" : {
        "type" : "object",
        "properties" : {
          "legalName" : {
            "type" : "string"
          },
          "registrationNumber" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "taxIdentifier" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "addressLine" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          }
        }
      },
      "ProspectLeadView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "businessName" : {
            "type" : "string"
          },
          "contactName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "email" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "sector" : {
            "type" : "string"
          },
          "city" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "weeklyVolume" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "note" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "source" : {
            "type" : [
              "string",
              "null"
            ]
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "contactedAt" : {
            "type" : [
              "string",
              "null"
            ],
            "format" : "date-time"
          }
        }
      },
...
      "TenantDirectoryEntry" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "contactEmail" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "organizerCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "TenantDirectoryPage" : {
        "type" : "object",
        "properties" : {
          "entries" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TenantDirectoryEntry"
            }
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}